+++ src/workprint/__init__.py
@@ -1,5 +1,13 @@
 """Workprint investigation engine."""
 
 from .engine import InvestigationEngine, InvestigationError
+from .models import Completeness, Observation, ObservationError, Reliability
 
-__all__ = ["InvestigationEngine", "InvestigationError"]
+__all__ = [
+    "Completeness",
+    "InvestigationEngine",
+    "InvestigationError",
+    "Observation",
+    "ObservationError",
+    "Reliability",
+]